backend/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
frozendict==2.4.6
gunicorn==23.0.0
html5lib==1.1
idna==3.10
joblib==1.4.2
lxml==5.3.0
multitasking==0.0.11
numpy==2.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
pycodestyle==2.12.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
Unipath==1.1
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.8.2
yfinance==0.2.54
